+++ VMWithEditor/UnitTests/State_Machine_Unit_Test/State_Machine_Unit_Test/internal_character_transition_unit_tests.h
@@ -1,7 +1,10 @@
 #ifndef INTERNAL_CHARACTER_TRANSITION_UNIT_TESTS_H
 #define INTERNAL_CHARACTER_TRANSITION_UNIT_TESTS_H
 
-bool unit_test_get_alpha_input_transition_character_type(const unsigned test_step);
+#include <stdbool.h>
+#include <stdint.h>
+
+bool unit_test_get_alpha_input_transition_character_type(const size_t test_step);
 bool unit_test_get_transition_character_type(const size_t test_step);
 
 #endif
